Add teacher classrooms cache
