+++ services/QuillLMS/spec/models/canvas_integration/teacher_classrooms_cache_spec.rb
@@ -2,6 +2,6 @@
 
 require 'rails_helper'
 
-RSpec.describe GoogleIntegration::TeacherClassroomsCache do
+RSpec.describe CanvasIntegration::TeacherClassroomsCache do
   it_behaves_like 'a teacher classrooms cache'
 end
